Fix compiler warnings on .NET 4.0 and above. Replace obsolete non-type-safe non-thread-safe Hashtable.Synchronized with type-safe thread-safe ConcurrentDictionary. Remove unnecessary casts and usings. Replace obsolete IPAddress.Address with IPAddress.GetAddressBytes in HostToDccInt. Remove unused _PassiveChannelSyncing code. Avoid double lookups in IrcClient._InterpretChannelMode.

+++ src/IrcFeatures/DccSend.cs
@@ -143,7 +143,7 @@ namespace Meebey.SmartIrc4net
                     
                     bytes = _File.Read(_Buffer, 0, _Buffer.Length);
                     try {
-                        Connection.GetStream().Write(_Buffer, 0, (int)bytes);
+                        Connection.GetStream().Write(_Buffer, 0, bytes);
                     } catch (IOException) {
                         bytes = 0;    // Connection Lost
                     }
